Simplify StealData returns; use int for len() results
TxtFileToHTML/SerializeBitmap rebuilt a Str from StealData()'s result,
which already carries the right pointer and length - just return it.
Also store len() results in int (its return type) instead of size_t.

+++ src/base/tests/Vec_ut.cpp
@@ -16,7 +16,7 @@ static size_t VecTestAppendFmt() {
         val = (val * 3) / 2; // somewhat exponential growth
         val += 15;
     }
-    size_t l = len(v);
+    int l = len(v);
     return l;
 }
